Move Base64 stuff to com
